Refactor: check transformations split (#748)
* fix: remove basicAuth if not explicitely added

* fix: remove empty basicAuth from checkeditornew test

* refactor: split out check transformation functions to their own check files

* refactor: remove selectableValues from form transformations
