+++ src/components/CheckEditor/FormComponents/DNSCheckResponseMatches.tsx
@@ -4,7 +4,7 @@ import { GrafanaTheme2, OrgRole } from '@grafana/data';
 import { Button, Checkbox, IconButton, Input, Label, Select, useStyles2 } from '@grafana/ui';
 import { css } from '@emotion/css';
 
-import { CheckFormValuesDns } from 'types';
+import { CheckFormValuesDns, ResponseMatchType } from 'types';
 import { hasRole } from 'utils';
 import { DNS_RESPONSE_MATCH_OPTIONS } from 'components/constants';
 
@@ -33,7 +33,7 @@ export const DNSCheckResponseMatches = () => {
                     name={`settings.dns.validations.${index}.responseMatch`}
                     rules={{ required: true }}
                     render={({ field }) => {
-                      const { ref, ...rest } = field;
+                      const { ref, onChange, ...rest } = field;
                       return (
                         <Select
                           {...rest}
@@ -41,6 +41,7 @@ export const DNSCheckResponseMatches = () => {
                           aria-label={`DNS Response Match ${userIndex}`}
                           options={DNS_RESPONSE_MATCH_OPTIONS}
                           invalid={Boolean(formState.errors.settings?.dns?.validations?.[index]?.responseMatch)}
+                          onChange={({ value }) => onChange(value)}
                         />
                       );
                     }}
@@ -81,7 +82,7 @@ export const DNSCheckResponseMatches = () => {
           variant="secondary"
           size="sm"
           disabled={!isEditor}
-          onClick={() => append({ matchType: DNS_RESPONSE_MATCH_OPTIONS[0], expression: '', inverted: false })}
+          onClick={() => append({ responseMatch: ResponseMatchType.Authority, expression: '', inverted: false })}
           data-fs-element="Add DNS response match button"
         >
           Add Regex Validation
