+++ src/components/CheckEditor/FormComponents/HttpCheckValidStatusCodes.tsx
@@ -25,8 +25,16 @@ export const HttpCheckValidStatusCodes = () => {
         control={control}
         name="settings.http.validStatusCodes"
         render={({ field }) => {
-          const { ref, ...rest } = field;
-          return <MultiSelect {...rest} options={validStatusCodes} disabled={!isEditor} inputId={id} />;
+          const { ref, onChange, ...rest } = field;
+          return (
+            <MultiSelect
+              {...rest}
+              options={validStatusCodes}
+              disabled={!isEditor}
+              inputId={id}
+              onChange={(values) => onChange(values.map((v) => v.value))}
+            />
+          );
         }}
       />
     </Field>
