+++ src/components/CheckEditor/FormComponents/RequestMethodSelect.tsx
@@ -15,7 +15,7 @@ type RequestMethodInputProps = {
 
 export const RequestMethodSelect = ({ 'aria-label': ariaLabel = `Request method`, name }: RequestMethodInputProps) => {
   const isEditor = hasRole(OrgRole.Editor);
-  const { formState } = useFormContext<CheckFormValues>();
+  const { control, formState } = useFormContext<CheckFormValues>();
   const errMessage = get(formState.errors, name);
 
   return (
@@ -28,9 +28,17 @@ export const RequestMethodSelect = ({ 'aria-label': ariaLabel = `Request method`
       data-fs-element="Check request method select"
     >
       <Controller
+        control={control}
         render={({ field }) => {
-          const { ref, ...rest } = field;
-          return <Select {...rest} options={METHOD_OPTIONS} aria-label={ariaLabel} />;
+          const { ref, onChange, ...rest } = field;
+          return (
+            <Select
+              {...rest}
+              options={METHOD_OPTIONS}
+              aria-label={ariaLabel}
+              onChange={({ value }) => onChange(value)}
+            />
+          );
         }}
         rules={{ required: true }}
         name={name}
